Rename createPaymentRequest template
